+++ upyog/imports.py
@@ -1,4 +1,5 @@
 import argparse
+import base64
 import enum
 import functools
 import importlib