+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 * Bugfixes
+* `image_to_base64` util to convert PIL -> base64 strings
+
 
 ## 0.7.7 -- 28 Aug 2023
 * Make pil interpolation map an internal attribute, as this API doesn't work with PIL-SIMD 9.0